+++ src/base/ApplicationEventBridgeRule.ts
@@ -2,16 +2,19 @@ import { Resource, TerraformResource } from 'cdktf';
 import { Construct } from 'constructs';
 import { eventbridge } from '@cdktf/provider-aws';
 
+export type Target = {
+  arn: string;
+  deadLetterArn?: string;
+  targetId?: string;
+  dependsOn: TerraformResource;
+};
+
 export interface ApplicationEventBridgeRuleProps {
   name: string;
   description?: string;
   eventBusName?: string;
   eventPattern: { [key: string]: any };
-  target?: {
-    arn: string;
-    targetId?: string;
-    dependsOn: TerraformResource;
-  };
+  targets?: Target[];
   tags?: { [key: string]: string };
 }
 
@@ -29,6 +32,10 @@ export class ApplicationEventBridgeRule extends Resource {
   }
 
   private createCloudwatchEventRule() {
+    if (this.config.targets?.length > 5) {
+      throw new Error('AWS allows only up to 5 targets per event bridge rule');
+    }
+    const eventBus = this.config.eventBusName ?? 'default';
     const rule = new eventbridge.CloudwatchEventRule(
       this,
       'event-bridge-rule',
@@ -36,17 +43,25 @@ export class ApplicationEventBridgeRule extends Resource {
         name: `${this.config.name}-Rule`,
         description: this.config.description,
         eventPattern: JSON.stringify(this.config.eventPattern),
-        eventBusName: this.config.eventBusName ?? 'default',
+        eventBusName: eventBus,
         tags: this.config.tags,
       }
     );
 
-    if (this.config.target) {
-      new eventbridge.CloudwatchEventTarget(this, 'event-bridge-target', {
-        rule: rule.name,
-        targetId: this.config.target.targetId,
-        arn: this.config.target.arn,
-        dependsOn: [this.config.target.dependsOn, rule],
+    if (this.config.targets) {
+      this.config.targets.forEach((t) => {
+        new eventbridge.CloudwatchEventTarget(
+          this,
+          `event-bridge-target-${t.targetId}`,
+          {
+            rule: rule.name,
+            targetId: t.targetId,
+            arn: t.arn,
+            deadLetterConfig: t.deadLetterArn ? { arn: t.deadLetterArn } : {},
+            dependsOn: [t.dependsOn, rule],
+            eventBusName: eventBus,
+          }
+        );
       });
     }